Web Development/cw/submission/requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.12.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
